==> 00_exploration.ipynb <==
This notebook is just for exploring the data

import pandas as pd
import numpy as np
import matplotlib.pyplot as plt

Data_path='C:\\Users\\gt\\End to End Data Science Project\\survey_results_public.csv'

#Read data and print the shape of it
#83439 response and 48 questions
raw_df=pd.read_csv(Data_path)
raw_df.shape

#display random answer
#Observations:Multiple answers needed to be splitted
raw_df.sample(1).iloc[0]

We can notice from here that:
- Some of the answers skipped like 'DatabaseHaveWorkedWith','WebframeWantToWorkWith',.....
- Some questions has multiple solutions and we need to split those answers
- the separator between words is ";" not ","

Our main focus in this data information related to the development not personal information and we should see mutliple responses to take an impression on the data and get more understanding

#print the general information to the dataframe and it's datatypes
raw_df.info()

Most of the datatypes are objects which means the data needs preprocessing so, that we can use it well

- some columns as Age, Age1stCode and YearsCodePro should have numerical data types not objects
- some columns as ResponseId is numerical and we will not make any calcualtions using it so, we can change it's datatype to object or remove the column

#Get stats for the numerical column
raw_df.describe()

CompTotal refers to current total compensation (salary, bonuses, and perks, before taxes and deductions)

#WhyYearsCodePro and YearsCode are not integers?
questionable_columns=['YearsCodePro','YearsCode']

for col in questionable_columns:
    print(col)
    print(raw_df[col].unique().tolist())
    print('------------------------')
    print()

it seems that there are options as 'More than 50 years' and 'Less than 1 year' and these options changed the datatype to object not an integer as expected so, we need to preprocess these columns before modeling
